+++ changelog.md
@@ -10,6 +10,7 @@
 * `[cellular]` Added a possibility to remove SIM PIN check on startup (SIM unlocking).
 * `[usb]` First implementation of USB's CDC and MTP protocols for communication with Mudita Center.
 * `[cellular]` Support for MMI Call Forwarding call.
+* `[cellular]` Added CLIR, CLIP, COLP, call waiting MMI support
 
 ### Changed